feat: Update login page text to use Portuguese language
feat: Add home page to the dashboard with balance, recent transactions, investment portfolio, and financial analysis components

feat: Add mode toggle component for theme switching

chore: Refactor AuthProvider to handle token management and logout functionality

chore: Update routes to include home page in the dashboard section

refactor: Update account query to use 'id' instead of 'userId'
